Add `<meta charset="utf-8">` to HTML files
The pages are being detected as `document.charset === "windows-1252"`. This causes corruption for unicode characters, e.g., "We’ve Got a Panic!" displays as "Weâ€™ve Got a Panic!" in `vsock.html` in my Firefox on macOS.

Adding an explicit meta tag is a reasonable fix. An alternative would be fixing the HTTP server to include the charset in the Content-Type header, but adding a meta tag is more explicit.